+++ src/views/AdminEditOffer.vue
@@ -85,6 +85,13 @@
                 outlined
               ></v-textarea>
 
+              <!-- Apply Contact Field -->
+              <v-text-field
+                v-model="offer.applyContact"
+                label="Usuario de Discord de contacto"
+                outlined
+              ></v-text-field>
+
               <!-- Apply Url Field -->
               <v-text-field
                 v-model="offer.applyUrl"