+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Pagination to offer feed in home page.
 - Offer requirements and about us can now contain markdown.
+- A lightbox with contact Discord user is opened when applying to offer that does not have apply link.
 
 ### Fixed